chessBackend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-chess>=1.999
pydantic>=2.0.0
openai>=0.27.0
